Add klims option to make_srcmap
